cloud_functions/data_load/requirements.txt
functions_framework==3.*
google-cloud-bigquery == 3.12.0